+++ evalml/tests/pipeline_tests/test_pipelines.py
@@ -2035,7 +2035,7 @@ def test_predict_has_input_target_name(
                     "date_index": "date",
                     "forecast_horizon": 2,
                 },
-                "Delayed Feature Transformer": {
+                "Time Series Featurizer": {
                     "gap": 0,
                     "max_delay": 0,
                     "forecast_horizon": 2,
@@ -2050,7 +2050,7 @@ def test_predict_has_input_target_name(
         clf = time_series_binary_classification_pipeline_class(
             parameters={
                 "Logistic Regression Classifier": {"n_jobs": 1},
-                "Delayed Feature Transformer": {
+                "Time Series Featurizer": {
                     "gap": 0,
                     "max_delay": 0,
                     "forecast_horizon": 2,
@@ -2071,7 +2071,7 @@ def test_predict_has_input_target_name(
         clf = time_series_multiclass_classification_pipeline_class(
             parameters={
                 "Logistic Regression Classifier": {"n_jobs": 1},
-                "Delayed Feature Transformer": {
+                "Time Series Featurizer": {
                     "gap": 0,
                     "max_delay": 0,
                     "forecast_horizon": 2,
@@ -2290,7 +2290,7 @@ def test_binary_pipeline_string_target_thresholding(
                 "date_index": "date",
                 "forecast_horizon": 3,
             },
-            "Delayed Feature Transformer": {"date_index": "date"},
+            "Time Series Featurizer": {"date_index": "date"},
         }
     )
     pipeline.fit(X_train, y_train)